feat: add agency entity and update bank

+++ src/main/java/com/wesleybertipaglia/bank/services/BankService.java
@@ -3,46 +3,55 @@ package com.wesleybertipaglia.bank.services;
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
+import java.util.stream.Collectors;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import com.wesleybertipaglia.bank.repositories.BankRepository;
+import com.wesleybertipaglia.bank.dtos.BankDTO;
 import com.wesleybertipaglia.bank.models.Bank;
+import com.wesleybertipaglia.bank.repositories.BankRepository;
 
 @Service
 public class BankService {
-
     @Autowired
     private BankRepository bankRepository;
 
-    public List<Bank> listBanks() {
-        return bankRepository.findAll();
+    public List<BankDTO> listBanks() {
+        return bankRepository.findAll().stream().map(this::convertToDTO).collect(Collectors.toList());
     }
 
-    public Optional<Bank> getBankById(UUID id) {
-        return bankRepository.findById(id);
+    public Optional<BankDTO> getBankById(UUID id) {
+        return bankRepository.findById(id).map(this::convertToDTO);
     }
 
-    public Bank createBank(Bank bank) {
-        return bankRepository.save(bank);
+    public BankDTO createBank(Bank bank) {
+        Bank savedBank = bankRepository.save(bank);
+        return convertToDTO(savedBank);
     }
 
-    public Optional<Bank> updateBank(UUID id, Bank bank) {
-        Optional<Bank> bankOptional = bankRepository.findById(id);
-
-        if (bankOptional.isPresent()) {
-            Bank bankToUpdate = bankOptional.get();
-            bankToUpdate.setName(bank.getName());
-            bankToUpdate.setCode(bank.getCode());
-
-            return Optional.of(bankRepository.save(bankToUpdate));
-        }
-
-        return Optional.empty();
+    public Optional<BankDTO> updateBank(UUID id, Bank bank) {
+        return bankRepository.findById(id).map(existingBank -> {
+            existingBank.setName(bank.getName());
+            existingBank.setCode(bank.getCode());
+            Bank updatedBank = bankRepository.save(existingBank);
+            return convertToDTO(updatedBank);
+        });
     }
 
     public void deleteBank(UUID id) {
-        bankRepository.deleteById(id);
+        if (bankRepository.existsById(id)) {
+            bankRepository.deleteById(id);
+        } else {
+            throw new IllegalArgumentException("Bank not found");
+        }
+    }
+
+    private BankDTO convertToDTO(Bank bank) {
+        BankDTO dto = new BankDTO();
+        dto.setId(bank.getId());
+        dto.setName(bank.getName());
+        dto.setCode(bank.getCode());
+        return dto;
     }
 }
